+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
 spring.application.name=Newsfeed-Team-Project
 
 jwt.secret.key=7JWI64WV7ZWY7IS47JqU67mE65GY6riw7J6F64uI64uk7K2I6r6466+47J6F64uI64uk7J6Y67aA7YOB65Oc66a964uI64uk
+jwt.access.time=1800000L
+jwt.refresh.time=1209600000L
